CORE-8608 Refactor AppDocEditView for dependency injection
